chore(contract-tests): adapt to the CE plugin module split
Main moved every CE connector to its own Go module under ce/plugins/
(#780), which the contract-test framework predates. Adapt:

- Move the shared contracttest helpers from the root module's
  internal/connectors/plugins/contracttest to pkg/domain/contracttest:
  plugin modules cannot import another module's internal/ tree, but all
  of them already depend on pkg/domain. Also proposed independently as
  PR #785 so it can merge and release ahead of this branch.
- Rewrite the contracttest import in all 11 consuming contract tests
  and the stale stripe client import in reversals_test.go.
- Add a TEMPORARY `replace github.com/formancehq/payments/pkg/domain
  => ../../../pkg/domain` to the 10 CE plugin modules whose contract
  tests use the helpers, so the package resolves before a pkg/domain
  release contains it. Drop these once PR #785 is released and the
  plugins bump their pkg/domain requirement.
- Point the contract-tests workflow's temporary pull_request paths at
  the ce/plugins/ locations.

The Justfile contract-tests target was already made module-aware during
the rebase (cd into ce/plugins/<c> or ee/plugins/<c> and test ./client).
Verified: all 14 connector contract suites compile and Skip green
without credentials; stripe/wise/powens/routable unit suites pass.

Constraint: CE plugin modules pin published pkg/domain versions; contracttest is not in any release yet
Rejected: per-plugin helper copies | 10+ duplicated files to keep in sync
Rejected: plugins require the root module | couples every plugin to the full payments dep tree
Confidence: high
Scope-risk: moderate
Directive: drop the per-plugin pkg/domain replaces once a pkg/domain release contains contracttest (PR #785)
Not-tested: contract runs with real credentials post-rebase
