+++ ce/plugins/adyen/client/contract_test.go
@@ -24,7 +24,7 @@ import (
 	"regexp"
 	"testing"
 
-	"github.com/formancehq/payments/internal/connectors/plugins/contracttest"
+	"github.com/formancehq/payments/pkg/domain/contracttest"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
