+++ ce/plugins/increase/client/contract_test.go
@@ -41,7 +41,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/formancehq/payments/internal/connectors/plugins/contracttest"
+	"github.com/formancehq/payments/pkg/domain/contracttest"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
