+++ ce/plugins/increase/go.mod
@@ -55,3 +55,5 @@ require (
 	golang.org/x/tools v0.45.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
+
+replace github.com/formancehq/payments/pkg/domain => ../../../pkg/domain
