+++ ce/plugins/qonto/client/contract_test.go
@@ -59,7 +59,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/formancehq/payments/internal/connectors/plugins/contracttest"
+	"github.com/formancehq/payments/pkg/domain/contracttest"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
