+++ ce/plugins/stripe/go.mod
@@ -52,3 +52,5 @@ require (
 	golang.org/x/text v0.38.0 // indirect
 	golang.org/x/tools v0.45.0 // indirect
 )
+
+replace github.com/formancehq/payments/pkg/domain => ../../../pkg/domain
